requirements.txt
# Core
pycocotools==2.0.11
requests==2.32.5
Pillow==12.0.0
matplotlib==3.7.0
numpy==1.26.4

# PyTorch
torch==2.10.0
torchvision==0.25.0

# SAM
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# ONNX
onnx>=1.15.0
onnxruntime>=1.17.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
